fix: review 三处修复（错误响应长度/事务泄漏/SGIP Trace cursor）
#1 [High] cmpp/sgip encode_error_response 的 total_len 推断为 usize，to_be_bytes() 误发 8 字节长度前缀 → 改 (total_len as u32)（畸形 PDU 错误响应路径，既有 bug）。

#2 [Med] TransactionManager on_report 匹配成功后清理 transactions(msg_id 键)+seq_to_msg_id，修复长生命周期客户端两表无界增长；同步更新单测断言（report 后条目应清空）。

#4 [Minor] SGIP message.rs Trace/TraceResp arm 改用 body 作用域 cursor，与其它 arm 一致、避免未来多 PDU 缓冲误解析。

验证：clippy 0 警告；connector/sgip/cmpp 单测全绿（含 1000 并发零漏匹配）；四协议集成 + transaction-manager 集成全绿。

+++ crates/rsms-connector/src/handlers/cmpp.rs
@@ -29,7 +29,7 @@ fn encode_error_response(command_id: u32, sequence_id: u32, status: CommandStatu
     let body_len = 4;
     let total_len = 12 + body_len;
     let mut pdu = Vec::with_capacity(total_len);
-    pdu.extend_from_slice(&total_len.to_be_bytes());
+    pdu.extend_from_slice(&(total_len as u32).to_be_bytes());
     pdu.extend_from_slice(&resp_command_id.to_be_bytes());
     pdu.extend_from_slice(&sequence_id.to_be_bytes());
     pdu.extend_from_slice(&status.as_u32().to_be_bytes());
